api/requirements.dev.txt
pytest==6.2.2
rope==0.18.0
flake8==3.8.4
autoflake==1.4
mypy==0.812
sqlalchemy_stubs==0.4
flake8_isort==4.0.0
isort==5.7.0
black==20.8b1
flake8-black==0.2.1
pytest-asyncio==0.14.0
